refactor tomorrow io ingestors
